added ability for vectors of vectors
The key is modifying the VECTOR_DEF and VECTOR()
so that the type concatenation will work for something
like VECTOR_DEF(VECTOR(long)).  To do this you need an
extra layer of indirection

    #define VECTOR(type) _VECTOR(type)
    #define _VECTOR(type) ...

This is because concatenation does not macro expand. Note
for it all to work we need vectors to be reference types.